requirements.txt
faker>=19.0.0
click>=8.0.0
pydantic>=2.0.0
kafka-python>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pyyaml>=6.0.0
